# server/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GSBS Server API",
    "version": "5.1.0",
    "description": "HTTP API for GSBS sync clients. Saves are pushed and pulled as whole files. Encrypted saves are AES-256-GCM ciphertext produced client-side; the server stores opaque bytes. All authenticated endpoints use a bearer token issued by /api/login."
  },
  "servers": [
    {
      "url": "/",
      "description": "The GSBS server host"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "totp_required": {
            "type": "boolean"
          },
          "totp_token": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limited",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/register": {
      "post": {
        "summary": "Register a new account (when open registration is enabled)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registered"
          },
          "403": {
            "description": "Registration disabled"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "Log in and register a device, returning a bearer token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "client_name": {
                    "type": "string"
                  },
                  "client_os": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token (or totp_required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/api/login/totp": {
      "post": {
        "summary": "Complete a two-factor login using the totp_token from /api/login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "totp_token": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  },
                  "client_name": {
                    "type": "string"
                  },
                  "client_os": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}